requirements.txt
fastapi
uvicorn
smolagents[litellm]
huggingface_hub
python-dotenv
gradio
openinference-instrumentation-smolagents
opentelemetry-sdk
opentelemetry-exporter-otlp
llama-cpp-python